requirements.txt
h5py==3.7.0
keras==2.10.0
numpy==1.23.3
Pillow==9.5.0
scipy==1.10.1
